# app/webroot/doc/openapi.yaml
openapi: 3.0.0
info:
  version: 2.4.148
  title: MISP Automation API
  description: |

    ### Getting Started
    Automation functionality is designed to automatically generate signatures for intrusion detection systems. 
    To enable signature generation for a given attribute, Signature field of this attribute must be set to Yes. 
    Note that not all attribute types are applicable for signature generation, currently we only support NIDS signature 
    generation for IP, domains, host names, user agents etc., and hash list generation for MD5/SHA1 values of file artefacts. 
    Support for more attribute types is planned. To make this functionality available for automated tools an authentication 
    key is used. This makes it easier for your tools to access the data without further form-based-authentication. 
    The [API](https://www.circl.lu/doc/misp/GLOSSARY.html#api) key can be found and managed under My Profile page (/users/view/me) 
    on a MISP instance.

    #### Accept and Content-Type headers 
    When performing your request, depending on the type of request, you might need to explicitly specify in what content 
    type you want to get your results. This is done by setting one of the below `Accept` headers:

        Accept: application/json
        Accept: application/xml

    When submitting data in a `POST`, `PUT` or `DELETE` operation you also need to specify in what content-type you encoded the payload. 
...

        Content-Type: application/json
...

...
    ```
...

...

...
servers:
...
paths:
...
  /admin/users:
    get:
      summary: "Get users"
      operationId: getUsers
      tags:
        - Users
      responses:
        "200":
          $ref: "#/components/responses/UserListResponse"
        "403":
          $ref: "#/components/responses/UnauthorizedApiErrorResponse"
        "404":
          $ref: "#/components/responses/NotFoundApiErrorResponse"
        default:
          $ref: "#/components/responses/ApiErrorResponse"
...
  /admin/logs:
    post:
      summary: "Get instance logs"
      operationId: getLogs
      tags:
        - Logs
      requestBody:
        $ref: "#/components/requestBodies/GetLogsRequest"
      responses:
        "200":
          $ref: "#/components/responses/GetLogsResponse"
        "403":
          $ref: "#/components/responses/UnauthorizedApiErrorResponse"
        "404":
          $ref: "#/components/responses/NotFoundApiErrorResponse"
        default:
          $ref: "#/components/responses/ApiErrorResponse"
...
components:
  schemas:
...
    UserId:
      type: string
      pattern: '^\d+$'
      maxLength: 10
      example: "12345"
...
    UserListItem:
      type: object
      properties:
        User:
          $ref: "#/components/schemas/User"
        Role:
          $ref: "#/components/schemas/Role"
        Organisation:
          type: object
          properties:
            id:
              $ref: "#/components/schemas/OrganisationId"
            name:
              $ref: "#/components/schemas/OrganisationName"
    UserList:
      type: array
      items:
        $ref: "#/components/schemas/UserListItem"

    # Organisations
...
    OrganisationName:
      type: string
      maxLength: 255
      example: "ORGNAME"
...
    LogActionType:
      type: string
      enum:
        - "accept"
        - "accept_delegation"
        - "acceptRegistrations"
        - "add"
        - "admin_email"
        - "attachTags"
        - "auth"
        - "auth_fail"
        - "blocklisted"
        - "captureRelations"
        - "change_pw"
        - "delete"
        - "disable"
        - "discard"
        - "discardRegistrations"
        - "edit"
        - "email"
        - "enable"
        - "enrichment"
        - "error"
        - "export"
        - "fetchEvent"
        - "file_upload"
        - "galaxy"
        - "include_formula"
        - "login"
        - "login_fail"
        - "logout"
        - "merge"
        - "pruneUpdateLogs"
        - "publish"
        - "publish_sightings"
        - "publish alert"
        - "pull"
        - "purge_events"
        - "push"
        - "registration"
        - "registration_error"
        - "remove_dead_workers"
        - "request"
        - "request_delegation"
        - "reset_auth_key"
        - "send_mail"
        - "security"
        - "serverSettingsEdit"
        - "tag"
        - "undelete"
        - "update"
        - "update_database"
        - "update_db_worker"
        - "upgrade_24"
        - "upload_sample"
        - "version_warning"
        - "warning"
        - "wipe_default"
    LogId:
      type: string
      pattern: '^\d+$'
      maxLength: 10
      example: "12345"
    LogTitle:
      type: string
      example: "Attribute (448272) from Event (1): Other/text foo"
    LogChange:
      type: string
      example: "name () => (ORGNAME)"
...
    Log:
      type: object
      properties:
        id:
          $ref: "#/components/schemas/LogId"
        title:
          $ref: "#/components/schemas/LogTitle"
        created:
          type: string
          format: datetime
        model:
          $ref: "#/components/schemas/ModelName"
        model_id:
          $ref: "#/components/schemas/ModelId"
        action:
          $ref: "#/components/schemas/LogActionType"
        user_id:
          $ref: "#/components/schemas/UserId"
        change:
          $ref: "#/components/schemas/LogChange"
        email:
          type: string
          format: email
        org:
          $ref: "#/components/schemas/OrganisationName"
        description:
          $ref: "#/components/schemas/LogDescription"
        ip:
          type: string
          example: "10.0.0.10"

    # General
...
    PageSearchFilter:
      type: integer
      format: int32
      nullable: true
      minimum: 0
    LimitSearchFilter:
      type: integer
      format: int32
      nullable: true
      minimum: 1
...
    DateRestSearchFilter:
      description: "You can use any of the valid time related filters (examples: 7d, timestamps, [14d, 7d] for ranges, etc.)"
      type: string
      nullable: true
    DateIntervalRestSearchFilter:
      description: "Interval described by two dates"
      type: array
      minimum: 2
      maximum: 2
      items:
        $ref: "#/components/schemas/DateRestSearchFilter"
...
    ModelId:
      type: string
      pattern: '^\d+$'
      maxLength: 10
      example: "12345"
    ModelName:
      type: string
      enum:
        - "AdminSetting"
        - "Allowedlist"
        - "AttachmentScan"
        - "Attribute"
        - "AttributeTag"
        - "AuditLog"
        - "AuthKey"
        - "Bruteforce"
        - "Cerebrate"
        - "Community"
        - "CorrelationExclusion"
        - "Correlation"
        - "Dashboard"
        - "DecayingModelMapping"
        - "DecayingModel"
        - "EventBlocklist"
        - "EventDelegation"
        - "EventGraph"
        - "EventLock"
        - "Event"
        - "EventReport"
        - "EventTag"
        - "FavouriteTag"
        - "Feed"
        - "FuzzyCorrelateSsdeep"
        - "GalaxyClusterBlocklist"
        - "GalaxyCluster"
        - "GalaxyClusterRelation"
        - "GalaxyClusterRelationTag"
        - "GalaxyElement"
        - "Galaxy"
        - "Inbox"
        - "Job"
        - "Log"
        - "MispObject"
        - "Module"
        - "News"
        - "NoticelistEntry"
        - "Noticelist"
        - "NotificationLog"
        - "ObjectReference"
        - "ObjectRelationship"
        - "ObjectTemplateElement"
        - "ObjectTemplate"
        - "Organisation"
        - "OrgBlocklist"
        - "Post"
        - "Regexp"
        - "RestClientHistory"
        - "Role"
        - "Server"
        - "ShadowAttribute"
        - "SharingGroupElement"
        - "SharingGroupOrg"
        - "SharingGroup"
        - "SharingGroupServer"
        - "SightingdbOrg"
        - "Sightingdb"
        - "Sighting"
        - "TagCollection"
        - "TagCollectionTag"
        - "Tag"
        - "Task"
        - "TaxonomyEntry"
        - "Taxonomy"
        - "TaxonomyPredicate"
        - "TemplateElementAttribute"
        - "TemplateElementFile"
        - "TemplateElement"
        - "TemplateElementText"
        - "Template"
        - "TemplateTag"
        - "Thread"
        - "ThreatLevel"
        - "User"
        - "UserSetting"
        - "WarninglistEntry"
        - "Warninglist"
        - "WarninglistType"
...
    ApiError:
      type: object
      required:
        - name
        - message
        - url
      properties:
        name:
          type: string
        message:
          type: string
        url:
          type: string
          example: "/attributes"
    UnauthorizedApiError:
      type: object
      required:
        - name
        - message
        - url
      properties:
        name:
          type: string
          example: "Authentication failed. Please make sure you pass the API key of an API enabled user along in the Authorization header."
        message:
          type: string
          example: "Authentication failed. Please make sure you pass the API key of an API enabled user along in the Authorization header."
        url:
          type: string
          example: "/attributes"
    NotFoundApiError:
      type: object
      required:
        - name
        - message
        - url
      properties:
        name:
          type: string
          example: "Invalid attribute"
        message:
          type: string
          example: "Invalid attribute"
        url:
          type: string
          example: "/attributes/1234"
...
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: Authorization
      description: |
        The authentication of the automation is performed via a secure key available in the MISP UI interface. 
        Make sure you keep that key secret as it gives access to the entire database! The [API](https://www.circl.lu/doc/misp/GLOSSARY.html#api) 
        key is available in the event actions menu under automation.

        The authorization is performed by using the following header in the API requests:

            Authorization: YOUR_API_KEY
            
        > **NOTE**: Since version 2.2 the usage of the authentication key in the URL is deprecated. 
          Instead, pass the auth key in an Authorization header in the request. The legacy option of having the auth key 
          in the URL is temporarily still supported but not recommended.
  requestBodies:
...
    GetLogsRequest:
      content:
        application/json:
          schema:
            type: object
            properties:
              page:
                $ref: "#/components/schemas/PageSearchFilter"
              limit:
                $ref: "#/components/schemas/LimitSearchFilter"
              id:
                $ref: "#/components/schemas/LogId"
              title:
                $ref: "#/components/schemas/LogTitle"
              created:
                oneOf:
                  - $ref: "#/components/schemas/DateRestSearchFilter"
                  - $ref: "#/components/schemas/DateIntervalRestSearchFilter"
              model:
                $ref: "#/components/schemas/ModelName"
              model_id:
                $ref: "#/components/schemas/ModelId"
              action:
                $ref: "#/components/schemas/LogActionType"
              user_id:
                $ref: "#/components/schemas/UserId"
              change:
                type: string
                description: "Text search for log change property"
                example: "%name () => (ORGNAME)%"
              email:
                type: string
                format: email
              org:
                type: string
                description: "Text search for log organisation name property"
                example: "ORG_%"
              description:
                type: string
                description: "Text search for log change property"
                example: "%updated by User%"
              ip:
                type: string
                description: "Text search for log ip property"
...
  responses:
...
    UserListResponse:
      description: "User list response"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/UserList"
...
    GetLogsResponse:
      description: "Get logs response"
      content:
        application/json:
          schema:
            type: array
            items:
              type: object
              properties:
                Log:
                  $ref: "#/components/schemas/Log"
...
    ApiErrorResponse:
      description: "Unexpected API error"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
    NotFoundApiErrorResponse:
      description: "The specified resource was not found"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/NotFoundApiError"
    UnauthorizedApiErrorResponse:
      description: "Authentication failed. Please make sure you pass the API key of an API enabled user along in the Authorization header."
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/UnauthorizedApiError"
...
security:
  - ApiKeyAuth: []